requirements.txt
aws-cdk-lib==2.55.1
constructs>=10.0.0,<11.0.0

# dev
black==22.12.0
isort==5.11.2
requests==2.28.1
